+++ src/lib/task-generators/mocha.ts
@@ -1,83 +1,45 @@
 import { Gulp } from "gulp";
-import gulpMocha from "gulp-mocha";
-import { Incident } from "incident";
-import { Minimatch } from "minimatch";
-import { toPosix } from "../project";
 import { AbsPosixPath } from "../types";
-import { asyncDone } from "../utils/async-done";
 import { TaskFunction } from "../utils/gulp-task-function";
-import * as matcher from "../utils/matcher";
-
-const mochaBin: AbsPosixPath = toPosix(require.resolve("mocha/bin/mocha"));
+import * as mocha from "../utils/mocha";
 
 export type MochaReporter = "spec";
 
 export interface MochaOptions {
+  rootDir: AbsPosixPath;
   testDir: AbsPosixPath;
   reporter?: MochaReporter;
+  /**
+   * Test `.spec.mjs` files instead of `.spec.js` (mixed is not supported)
+   * Default: `false`
+   */
+  mjs?: boolean;
+  colors?: boolean;
 }
 
-interface ResolvedMochaOptions {
+export interface ResolvedMochaOptions {
+  rootDir: AbsPosixPath;
   testDir: AbsPosixPath;
   reporter: MochaReporter;
+  mjs: boolean;
+  colors: boolean;
 }
 
-function resolveOptions(options: MochaOptions): ResolvedMochaOptions {
-  return {reporter: "spec", ...options};
-}
-
-export interface Sources {
-  baseDir: string;
-  specs: string[];
-}
-
-export function getSources(options: MochaOptions): Sources {
-  const baseDir: string = options.testDir;
-  const specs: string = matcher.asString(matcher.join(baseDir, new Minimatch("**/*.spec.js")));
-
-  return {
-    baseDir,
-    specs: [specs],
-  };
-}
-
-/**
- * Return the command line arguments equivalent to this task
- *
- * @param options Mocha options
- * @param colors Force colors
- * @return Command line arguments
- */
-export function getCommand(options: MochaOptions, colors: boolean): string[] {
-  const resolved: ResolvedMochaOptions = resolveOptions(options);
-  const sources: Sources = getSources(options);
-  const result: string[] = [mochaBin];
-  result.push("--ui", "bdd");
-  result.push("--reporter", resolved.reporter);
-  if (colors) {
-    result.push("--colors");
-  }
-  result.push("--", ...sources.specs);
-  return result;
+export function resolveMochaOptions(options: MochaOptions): ResolvedMochaOptions {
+  return {reporter: "spec", mjs: false, colors: true, ...options};
 }
 
 export function generateTask(gulp: Gulp, options: MochaOptions): TaskFunction {
-  const resolved: ResolvedMochaOptions = resolveOptions(options);
-  const sources: Sources = getSources(options);
+  const resolved: ResolvedMochaOptions = resolveMochaOptions(options);
 
-  const task: TaskFunction = async function (): Promise<any> {
-    return asyncDone(() => {
-      gulp
-        .src(sources.specs, {base: sources.baseDir})
-        .pipe(gulpMocha({
-          reporter: resolved.reporter,
-        }));
-    })
-      .catch((err: Error) => {
-        // Swallow original error because its error message is too long and just add noise
-        // TODO: Differentiate failed test error from runtime error
-        throw Incident("MochaError");
-      });
+  const task: TaskFunction = async function (): Promise<void> {
+    return mocha.run({
+      cwd: resolved.rootDir,
+      testDir: resolved.testDir,
+      reporter: resolved.reporter,
+      colors: true,
+      mjs: resolved.mjs,
+    });
   };
   task.displayName = getTaskName();
 
@@ -85,5 +47,5 @@ export function generateTask(gulp: Gulp, options: MochaOptions): TaskFunction {
 }
 
 export function getTaskName(): string {
-  return "_mocha:run";
+  return "_:mocha:run";
 }